move load_compressed_proof_from_bzip2
