+++ crates/raito-spv-client/src/verify.rs
@@ -3,17 +3,17 @@
 
 use bitcoin::Network;
 use bitcoin::{block::Header as BlockHeader, consensus, MerkleBlock, Transaction};
-use bzip2::read::BzDecoder;
 use cairo_air::utils::{get_verification_output, VerificationOutput};
 use cairo_air::{CairoProof, PreProcessedTraceVariant};
 use raito_spv_core::block_mmr::{BlockInclusionProof, BlockMMR};
-use std::{io::Read, path::PathBuf};
+use std::path::PathBuf;
 use stwo_prover::core::vcs::blake2_merkle::{Blake2sMerkleChannel, Blake2sMerkleHasher};
 use tracing::info;
 
 use crate::format::format_transaction;
 use crate::proof::{BootloaderOutput, ChainState, CompressedSpvProof, TaskResult};
 use crate::work::verify_subchain_work;
+use crate::fetch::load_compressed_proof_from_bzip2;
 
 /// CLI arguments for the `verify` subcommand
 #[derive(Clone, Debug, clap::Args)]
@@ -50,39 +50,6 @@ impl Default for VerifierConfig {
             task_output_size: 8,
         }
     }
-}
-
-/// Load a compressed proof from disk that was saved using bincode binary codec with bzip2 compression
-///
-/// - `proof_path`: Path to the bzip2 compressed proof file
-///
-/// This function first decompresses the bzip2 file, then deserializes the bytes
-/// using bincode binary codec, providing the symmetric operation to
-/// `save_compressed_proof_with_bzip2`.
-pub fn load_compressed_proof_from_bzip2(
-    proof_path: &PathBuf,
-) -> Result<CompressedSpvProof, anyhow::Error> {
-    info!(
-        "Loading and decompressing proof from {}",
-        proof_path.display()
-    );
-
-    // Step 1: Read and decompress the file
-    let file = std::fs::File::open(proof_path)?;
-    let mut bz_decoder = BzDecoder::new(file);
-    let mut decompressed_bytes = Vec::new();
-    bz_decoder.read_to_end(&mut decompressed_bytes)?;
-
-    info!(
-        "Decompressed {} bytes, now deserializing...",
-        decompressed_bytes.len()
-    );
-
-    // Step 2: Deserialize the decompressed bytes using bincode
-    let proof: CompressedSpvProof = bincode::deserialize(&decompressed_bytes)?;
-
-    info!("Successfully loaded compressed proof");
-    Ok(proof)
 }
 
 /// Run the `verify` subcommand: read a proof from disk and verify it
